move member-row into members dir
